+++ src/App.js
@@ -1,20 +1,18 @@
 // @flow
-
-import React, { Component } from 'react'
-import ReactDom from 'react-dom'
-import { Route, Switch,  } from 'react-router-dom'
-import { Provider } from 'react-redux'
-import { ConnectedRouter } from 'connected-react-router'
-import history from './store/history'
-import store from './store'
-import './styles/reset'
-import coreInstance from './coreInstance'
-import {css} from 'react-emotion'
-import NoComponent from './components/NoComponent'
-import Menu from "./components/Menu";
-import { Header } from './components/Header'
+import React, { Component } from 'react';
+import { Route, Switch } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import { ConnectedRouter } from 'connected-react-router';
+import history from './store/history';
+import store from './store';
+import './styles/reset';
+import coreInstance from './coreInstance';
+import { css } from 'react-emotion';
+import NoComponent from './components/NoComponent';
+import Menu from './components/Menu';
+import Header from './components/Header';
 // $FlowFixMe
-import 'antd/dist/antd.less'
+import 'antd/dist/antd.less';
 
 const sideColor = '#343434;';
 const styles = {
@@ -23,12 +21,6 @@ const styles = {
     margin: 0 auto;
     height: 100vh;
     flex-direction: column;
-	`,
-  head: css`
-		height: 50px;
-		background: #61000D;
-		flex-grow: 0;
-    flex-shrink: 0;
 	`,
   main: css`
 		display:flex;
@@ -63,40 +55,39 @@ const styles = {
 };
 
 const mapRoutesModule = () => {
-  const modules = coreInstance.getModules()
-  return modules.map(module => <Route key={module.namespace} path={'/'+module.namespace} component={module.RootComponent}/>)
+  const modules = coreInstance.getModules();
+  return modules.map(module => (
+    <Route key={module.namespace} path={'/' + module.namespace} component={module.RootComponent}/>
+  ));
 };
 
 export default class App extends Component<any> {
   componentDidMount() {
     coreInstance.subscribe('registerModule', () => {
-      this.forceUpdate()
-    })
+      this.forceUpdate();
+    });
   }
   render() {
     return (
       <Provider store={store}>
         <div className={styles.layout}>
-          <div className={styles.head}>
-            <Header/>
-          </div>
+          <Header/>
           <div className={styles.main}>
 
-              <div className={styles.sidemenu}>
-                <Menu/>
-              </div>
-              <div className={styles.content}>
-                <ConnectedRouter history={history}>
-                    <Switch >
-                      {mapRoutesModule()}
-                      <Route path={'/'} component={NoComponent}/>
-                    </Switch>
-                </ConnectedRouter>
-              </div>
-
+            <div className={styles.sidemenu}>
+              <Menu/>
+            </div>
+            <div className={styles.content}>
+              <ConnectedRouter history={history}>
+                <Switch >
+                  {mapRoutesModule()}
+                  <Route path={'/'} component={NoComponent}/>
+                </Switch>
+              </ConnectedRouter>
+            </div>
           </div>
         </div>
       </Provider>
-    )
+    );
   }
 }