+++ src/coreInstance.js
@@ -1,5 +1,3 @@
 import Core from './core'
 
-const coreInstance = new Core()
-window.tarantool_enterprise_core = coreInstance
-export default coreInstance
+export default new Core()