+++ src/store/constants.js
@@ -1,4 +1,9 @@
+// @flow
 export const SELECT_MENU = 'SELECT_MENU'
 export const UPDATE_LINK = 'UPDATE_LINK'
 export const RESET = 'RESET'
+
+export const TITLE_SET = 'TITLE_SET';
+export const TITLE_RESET = 'TITLE_RESET';
+
 export const LOCATION_CHANGE = '@@router/LOCATION_CHANGE';